Issue #42 Outline error list
Add fragment and layout for showing error lists

Move fragments into own packages

+++ client-core/src/main/java/de/njsm/stocks/client/business/entities/ErrorDescription.java
@@ -36,7 +36,7 @@ public abstract class ErrorDescription {
 
     public abstract ErrorDetails errorDetails();
 
-    public static ErrorDescription create(long id, StatusCode statusCode, String stackTrace, String errorMessage, ErrorDetails errorDetails) {
+    public static ErrorDescription create(int id, StatusCode statusCode, String stackTrace, String errorMessage, ErrorDetails errorDetails) {
         return new AutoValue_ErrorDescription(id, statusCode, stackTrace, errorMessage, errorDetails);
     }
 }
